feat: add harness-based jest runner (#4)
Catching up with Jest, which has been under development for years, isn't easy (especially for a team with only a single developer). Therefore, it may make more sense to adopt Jest and use it as a CLI for Harness. This pull request implements a custom Jest runner based on Harness, capable of running tests on emulators, just like the Harness CLI.

With this approach, there's no need to maintain a separate CLI, and we can leverage the existing infrastructure for reports and other integrations. For example, I was able to set up test coverage support in just a few minutes, rather than hours.