feat: add 'resetEnvironmentBetweenTestFiles' config property (#11)
This pull request adds a new configuration property that allows you to opt out of test file-based isolation and run multiple test files in a single JavaScript environment. This can speed up test execution at the cost of isolation, so if your tests modify the global context in any way, you should not enable it.