feat: throw when jest globals are used (#12)
Users might forget to replace all instances of the jest global with harness. We should remind them to do so as soon as we detect any usage. This pull request introduces the necessary changes to throw an error whenever the jest global is used or when @jest/globals is imported.