feat: run Metro internally (#27)
Currently, Harness runs Metro through the React Native CLI. This means the tool has no direct control over the configuration and relies on CLI parameters, as well as the user, to wrap the config with the withRnHarness function. By internalizing Metro and leveraging its JS API, Harness can dynamically augment the configuration on the fly. As a result, users no longer need to manually update the config file, and this approach also enables further optimizations.